UnitSlots for Unit data, properly keeping track data, sending data
Who knew that referencing your data to a prefab to keep track of individual data for various objects is a terrible, horrible idea? Also, learned about the factory principle! I don't think it's particularly useful for this project. But I now know about it and how to use it! Isn't that neat?