Implement most of Haven Town UI and test all units (!!!) Also more mines
Creature dwelling build is still not consistently bug-free: the building data SO's produced unit still gets set to null somewhere, so more abstraction is required to get the system working properly.

At least the new units seem to be working fine, and making the prefab be held in UnitStats did allow for proper addition of a unit not previously present in the army with proper combat initiation, which is good.

When a unit dies on an army, the next unit in that army's turn being skipped is still an issue. Need to fix

Also reminder to see what's causing the 1/1/1 movement slider to not update when hero moves.